Change types for memory management functions

+++ data_structures/allocator.h
@@ -9,8 +9,8 @@ typedef struct {
 
 bitmap_t get_allocator_bitmap (allocator_t * allocator);
 
-void set_alloc_bitmap (allocator_t * allocator, bitmap_t loc, int length);
+void set_alloc_bitmap (allocator_t * allocator, bitmap_t loc, uint32_t length);
 
-int allocate (allocator_t * allocator, int n);
+uint32_t allocate (allocator_t * allocator, size_t n);
 
-int allocator_free(allocator_t * allocator, uint32_t index, int size);
+int allocator_free(allocator_t * allocator, uint32_t index, size_t size);
